cleanup, comment out broken svelteConfig import
